v0.2.0(coordijk): widen inner arithmetic + testCoordIjkExtRes (CO-G1..G6)
Third v0.2.0 widening per POC-5. The struct typedef (Session-5) and
the maxDim/unitScale tables (commit f84ab171) are necessary but not
sufficient — the CoordIJK arithmetic functions in coordijk.h hold int
locals that capture the now-int64 fields and silently narrow. UBSAN
trips at coordijk.h:487-505 (_upAp7/_upAp7r) when running res 21+
round-trips. POC-5 §"CO-19..22" predicted exactly these sites.

Widening:
- coordijk.h all 4 _upAp7/_upAp7r/_upAp7Checked/_upAp7rChecked: int
  i/j locals → int64_t; (int)lround → (int64_t)lround; explicit
  (double) casts on the floating arithmetic for clarity.
- coordijk.h _ijkNormalize: int min → int64_t.
- coordijk.h _ijkNormalizeCouldOverflow: int max,min → int64_t;
  ADD/SUB_INT32S_OVERFLOWS replaced with __builtin_*_overflow
  (type-generic, exact int64).
- coordijk.h _ijkToHex2d: int i,j → int64_t; (double) casts on
  the Vec2d arithmetic.
- coordijk.h _hex2dToCoordIJK: int m1,m2 → int64_t; (int)x1 →
  (int64_t)x1; long long int axisi/diff → int64_t; (int)cast on
  the final assignment widened.
- coordijk.h INT32_MAX_3 → INT64_MAX_3. With int64 fields the guard
  is effectively unreachable (worst H3 ext magnitude ≈ 2e9, well
  below INT64_MAX_3 ≈ 3e18) but kept for defense.

New test (CO-G1..G6, ~3000 assertions):
- src/apps/testapps/testCoordIjkExtRes.c — 6 TEST blocks:
  * CO-G1: 500 res 20 (Class II) cell-center round-trips bit-identical.
  * CO-G2: 500 res 21 (Class III) round-trips bit-identical (THE
    headline gate — pre-widening: 247/500; post-widening: 500/500).
  * CO-G3: 500 res 22 (Class II) round-trips bit-identical.
  * CO-G4: pentagon res 19+ getPentagons returns 12 distinct cells
    without UBSAN (POC-5 §"CO-24" boundary cleared).
  * CO-G5: stock res 0-15 byte-identity guard (CoordIJK widening
    must not regress stock).
  * CO-G6: 3 POC-5 user-spec geos × res 20-22 round-trip.

Stock test contract update (per CLAUDE.md non-negotiable #2):
- testCoordIjkInternal.c:54-94 — INT32_MAX → INT64_MAX in the
  _upAp7Checked / _upAp7rChecked overflow tests. The test intent
  ("i + i overflows", "i * 3 overflows", etc.) is preserved at the
  new int64 bound.

Local ctest: **327/327 PASS** under ASAN+UBSAN (326 prior + new
testCoordIjkExtRes). Pentagon path at res 19+, hexagon path at res
20-22, all 3 POC-5 geos: all clean.

POC-5's predicted boundary is now the post-widening pass region.

+++ src/apps/testapps/testCoordIjkInternal.c
@@ -51,21 +51,25 @@ SUITE(coordIjkInternal) {
         t_assert(_ijkMatches(&ijk, &i), "Invalid neighbor is self");
     }
 
+    // H3-EXTENDED: contract update per v0.2.0 — CoordIJK widened int → int64.
+    // Overflow guards trigger at int64 limits, not int32. INT32_MAX inputs no
+    // longer overflow; INT64_MAX inputs do. Stock test intent (i + i / i * 3
+    // / j + j / etc. should overflow) preserved at the new bound.
     TEST(_upAp7Checked) {
         CoordIJK ijk;
 
         _setIJK(&ijk, 0, 0, 0);
         t_assertSuccess(_upAp7Checked(&ijk));
-        _setIJK(&ijk, INT32_MAX, 0, 0);
+        _setIJK(&ijk, INT64_MAX, 0, 0);
         t_assert(_upAp7Checked(&ijk) == E_FAILED, "i + i overflows");
-        _setIJK(&ijk, INT32_MAX / 2, 0, 0);
+        _setIJK(&ijk, INT64_MAX / 2, 0, 0);
         t_assert(_upAp7Checked(&ijk) == E_FAILED, "i * 3 overflows");
-        _setIJK(&ijk, 0, INT32_MAX, 0);
+        _setIJK(&ijk, 0, INT64_MAX, 0);
         t_assert(_upAp7Checked(&ijk) == E_FAILED, "j + j overflows");
         // This input should be invalid because j < 0
-        _setIJK(&ijk, INT32_MAX / 3, -2, 0);
+        _setIJK(&ijk, INT64_MAX / 3, -2, 0);
         t_assert(_upAp7Checked(&ijk) == E_FAILED, "(i * 3) - j overflows");
-        _setIJK(&ijk, INT32_MAX / 3, INT32_MAX / 2, 0);
+        _setIJK(&ijk, INT64_MAX / 3, INT64_MAX / 2, 0);
         t_assert(_upAp7Checked(&ijk) == E_FAILED, "i + (j * 2) overflows");
         // This input should be invalid because j < 0
         _setIJK(&ijk, -1, 0, 0);
@@ -77,16 +81,16 @@ SUITE(coordIjkInternal) {
 
         _setIJK(&ijk, 0, 0, 0);
         t_assertSuccess(_upAp7rChecked(&ijk));
-        _setIJK(&ijk, INT32_MAX, 0, 0);
+        _setIJK(&ijk, INT64_MAX, 0, 0);
         t_assert(_upAp7rChecked(&ijk) == E_FAILED, "i + i overflows");
-        _setIJK(&ijk, 0, INT32_MAX, 0);
+        _setIJK(&ijk, 0, INT64_MAX, 0);
         t_assert(_upAp7rChecked(&ijk) == E_FAILED, "j + j overflows");
-        _setIJK(&ijk, 0, INT32_MAX / 2, 0);
+        _setIJK(&ijk, 0, INT64_MAX / 2, 0);
         t_assert(_upAp7rChecked(&ijk) == E_FAILED, "3 * j overflows");
-        _setIJK(&ijk, INT32_MAX / 2, INT32_MAX / 3, 0);
+        _setIJK(&ijk, INT64_MAX / 2, INT64_MAX / 3, 0);
         t_assert(_upAp7rChecked(&ijk) == E_FAILED, "(i * 2) + j overflows");
         // This input should be invalid because i < 0
-        _setIJK(&ijk, -2, INT32_MAX / 3, 0);
+        _setIJK(&ijk, -2, INT64_MAX / 3, 0);
         t_assert(_upAp7rChecked(&ijk) == E_FAILED, "(j * 3) - 1 overflows");
         // This input should be invalid because j < 0
         _setIJK(&ijk, -1, 0, 0);
